+++ src/zeroband/checkpointingv2/__init__.py
@@ -1,0 +1,23 @@
+from typing import Any
+from dataclasses import dataclass
+from torch.distributed.checkpoint.stateful import Stateful
+
+from .simple_ckpt_manager import SimpleCkptManager
+
+
+@dataclass
+class TrainingProgress(Stateful):
+    total_tokens: int
+    outer_step: int
+    step: int
+
+    def state_dict(self) -> dict[str, Any]:
+        return {"total_tokens": self.total_tokens, "outer_step": self.outer_step, "step": self.step}
+
+    def load_state_dict(self, state_dict: dict[str, Any]) -> None:
+        self.total_tokens = state_dict["total_tokens"]
+        self.outer_step = state_dict["outer_step"]
+        self.step = state_dict["step"]
+
+
+__all__ = ["SimpleCkptManager", "TrainingProgress"]
