+++ src/zeroband/checkpointingv2/simple_ckpt_manager.py
@@ -1,5 +1,4 @@
-from typing import Iterable, Optional, List, Any
-from dataclasses import dataclass
+from typing import Iterable, Optional, List
 import torch
 from torch.distributed.checkpoint.stateful import Stateful
 from torch.distributed._tensor.api import DTensor
@@ -10,21 +9,6 @@ import multiprocessing as mp
 from multiprocessing.synchronize import Event as EventType
 from safetensors import safe_open
 from safetensors.torch import save_file
-
-
-@dataclass
-class TrainingProgress(Stateful):
-    total_tokens: int
-    outer_step: int
-    step: int
-
-    def state_dict(self) -> dict[str, Any]:
-        return {"total_tokens": self.total_tokens, "outer_step": self.outer_step, "step": self.step}
-
-    def load_state_dict(self, state_dict: dict[str, Any]) -> None:
-        self.total_tokens = state_dict["total_tokens"]
-        self.outer_step = state_dict["outer_step"]
-        self.step = state_dict["step"]
 
 
 def _is_path(path_or_url: str) -> bool:
